fix: ensure pontos does not stop on node error (#183)
## Description

As seen today with @remiroyc, Juno node and the gateway are not able to
process some call to old / bad contracts.

In our case, this causes Juno to crash, and the docker container is
restarted.
To ensure that we can follow our operations in this case, this PR
ensures that
the main loop of Pontos is not stopping in such situation.

## What type of PR is this? (check all applicable)

- [X] 🐛 Bug Fix (`fix:`)